+++ tests/models/methods.js
@@ -199,7 +199,7 @@ Tinytest.addAsync('Models - Method - Waited On - track wait time of queued messa
 
   await sleep(1000);
 
-  const metrics = Kadira.models.methods._getMetrics(Ntp._now(), methodId);
+  const metrics = findMetricsForMethod(methodId);
 
   test.isTrue(metrics.waitedOn > 25, `${metrics.waitedOn} should be greater than 25`);
   test.isTrue(metrics.waitedOn <= 6000, `${metrics.waitedOn} should be less than 6k`);
@@ -222,7 +222,7 @@ Tinytest.addAsync('Models - Method - Waited On - check unblock time', async (tes
 
   await sleep(1000);
 
-  const metrics = Kadira.models.methods._getMetrics(Ntp._now(), methodId);
+  const metrics = findMetricsForMethod(methodId);
 
   test.isTrue(metrics.waitedOn <= 1, 'waitedOn should be less or equal than 1');
 
@@ -242,7 +242,7 @@ Tinytest.addAsync('Models - Method - Waited On - track wait time of next message
 
   await sleep(200);
 
-  const metrics = Kadira.models.methods._getMetrics(Ntp._now(), slowMethod);
+  const metrics = findMetricsForMethod(slowMethod);
   test.isTrue(metrics.waitedOn >= 20, `${metrics.waitedOn} should be greater than 20`);
 
   done();