+++ CHANGELOG.md
@@ -5,6 +5,19 @@
 * Support Meteor 3
 * Add `Monti.event`
 * Tracer supports parallel events
+
+## v2.48.0
+March 6, 2024
+
+* Track waitedOn time for methods and publications - the amount of time it is blocking other methods and publications
+* Fix using outdated staticFilesByArch in development on some Meteor versions
+* Add `Monti.event` to match v3 of agent
+* Fix tracking wait time for publications
+* Run object filters on HTTP req body
+* Document Monti.tracer.redactField
+* Rounds event loop lag to nearest 500 microseconds
+* Remove unused pctEvloopBlock metric
+
 ## 2.47.4
 Nov 24, 2023